+++ tests/testthat/test_cv.R
@@ -10,6 +10,14 @@ test_that("Test CV returns", {
   expect_equal(dim(res2$multilabel), c(3, length(res1)))
   expect_named(res2$labels, rownames(toyml$labels))
   expect_equal(nrow(res2$labels[[1]]), 3)
+
+  res3 <- cv(toyml, br, base.algorithm="RANDOM", cv.folds=3,
+             cv.predictions=TRUE)
+  expect_length(res3, 2)
+  expect_named(res3, c("multilabel", "predictions"))
+  expect_equal(dim(res3$multilabel), c(3, length(res1)))
+  expect_length(res3$predictions, 3)
+  expect_equal(colnames(res3$predictions[[1]]), names(res2$labels))
 })
 
 test_that("Test CV seed", {